changed variable type for last_digit back to int

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -10,7 +10,7 @@
 
 int print_last_digit(int n)
 {
-	long long int last_digit;
+	int last_digit;
 
 	if (n > 0 || n == 0)
 	{
